refactor: AssetInfoの定義をsharedへ集約
