+++ lambda/web-scraping/src/domain/__init__.py
@@ -1,12 +1,13 @@
 """Domain レイヤー: モデル、インターフェース、例外"""
 
+from shared.domain.asset_object import DcpAssetInfo
+
 from .artifact_interface import IArtifactRepository
 from .exceptions import (
     ArtifactUploadError,
     ScrapingFailed,
     WebScrapingFailed,
 )
-from .extraction_object import DcpAssetInfo
 from .scraping_interface import IScraper
 from .scraping_object import ScrapingParams
